Bug 1689620 - Move outcome checker inside auto_perf_sheriffing package
