+++ treeherder/perf/auto_perf_sheriffing/secretary_tool.py
@@ -6,7 +6,7 @@ from django.conf import settings as django_settings
 
 from treeherder.perf.models import BackfillRecord, BackfillReport, PerformanceSettings
 from treeherder.utils import default_serializer
-from treeherder.perf.outcome_checker import OutcomeChecker, OutcomeStatus
+from treeherder.perf.auto_perf_sheriffing.outcome_checker import OutcomeChecker, OutcomeStatus
 
 logger = logging.getLogger(__name__)
